# docs/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pi Docs API
  description: |
    REST API for Pi Docs - A document management system.
    
    This API can be used by web frontends, mobile apps, and third-party integrations.
    
    ## Authentication
    The API uses Bearer Token (JWT) authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
    
    ## Base URL
    - Production: `https://api.yourdomain.com/api`
    - Development: `http://localhost:3001/api`
  version: 1.0.0
  contact:
    name: Pi Docs Support
    url: https://github.com/yourusername/pi-docs
servers:
  - url: https://api.yourdomain.com/api
    description: Production server
  - url: http://localhost:3001/api
    description: Development server
...
paths:
...
  /documents:
    get:
      tags:
        - Documents
      summary: List documents
      description: Get all documents for the authenticated user
      operationId: listDocuments
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: query
          description: Filter by category ID
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          description: Search in title and description
          schema:
            type: string
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
    post:
      tags:
        - Documents
      summary: Create document
      description: Create a new document (with optional file upload)
      operationId: createDocument
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                description:
                  type: string
                categoryId:
                  type: string
                  format: uuid
                issuedAt:
                  type: string
                  format: date-time
                expiresAt:
                  type: string
                  format: date-time
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Document created
          content:
            application/json:
              schema:
                type: object
                properties:
                  document:
                    $ref: '#/components/schemas/Document'
  /documents/{id}:
    get:
      tags:
        - Documents
      summary: Get document
      description: Get a specific document by ID
      operationId: getDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                type: object
                properties:
                  document:
                    $ref: '#/components/schemas/Document'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags:
        - Documents
      summary: Update document
      description: Update a document
      operationId: updateDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocumentRequest'
      responses:
        '200':
          description: Document updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  document:
                    $ref: '#/components/schemas/Document'
    delete:
      tags:
        - Documents
      summary: Delete document
      description: Delete a document and its associated file
      operationId: deleteDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document deleted
  /documents/{id}/download:
    get:
      tags:
        - Documents
      summary: Download document file
      description: Download the file associated with a document
      operationId: downloadDocument
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          description: Document or file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - ACTIVE
            - EXPIRING
            - EXPIRED
        issuedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        fileName:
          type: string
          nullable: true
        fileMime:
          type: string
          nullable: true
        fileSize:
          type: integer
          nullable: true
        storageProvider:
          type: string
          enum:
            - LOCAL
            - S3
        storageKey:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - status
        - storageProvider
        - createdAt
        - updatedAt
...
    UpdateDocumentRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        categoryId:
          type: string
          format: uuid
          nullable: true
        issuedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
          required:
            - code
            - message
      required:
        - error
security:
  - bearerAuth: []